feat(client-server): Finalization of CRD operation - Replacement of notification snackbar - Automatic update of notes on the page

+++ server/routes/note.js
@@ -7,6 +7,7 @@ const prisma = new PrismaClient()
 router.get('/', async (req, res) => {
   const notes = await prisma.note.findMany({
     take: 50,
+    orderBy: { date_created: 'desc' },
   })
   res.json(notes)
 })
@@ -14,22 +15,27 @@ router.get('/', async (req, res) => {
 router.post(`/`, async (req, res) => {
   try{
     const result = await prisma.note.create({
-      data: { ...req.body, date_created: new Date() },
+      data: { title: req.body.title, content: req.body.content, date_created: req.body.date_created || new Date() },
     })
     res.json(result)  
   }
   catch(err){
-    res.json(err, 500)
+    res.status(500).json(err)
   }
 })
 
 router.delete(`/:id`, async (req, res) => {
   try{
-    const { id } = req.params
-    const post = await prisma.note.delete({
-      where: { id: Number(id) },
-    })
-    res.json(post)
+    const id = parseInt(req.params.id)
+    if (typeof id == 'number' && id > 0) {
+      const post = await prisma.note.delete({
+        where: { id: Number(id) },
+      })
+      res.json(post)
+    }
+    else {
+      return res.status(400).json({ error: 'Invalid id' })
+    }
   }
   catch(err){
     res.status(500).json(err)
